% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{collapse}
\alias{collapse}
\title{Collapse a character vector}
\usage{
collapse(x, sep = "", width = Inf, last = "")
}
\arguments{
\item{x}{The character vector to collapse.}

\item{sep}{a character string to separate the terms.  Not
    \code{\link{NA_character_}}.}

\item{width}{The maximum string width before truncating with \sQuote{...}.}

\item{last}{String used to separate the last two items if \sQuote{x} has at least
2 items.}
}
\description{
Collapses a character vector of any length into a length 1 vector.
}
\examples{
collapse(glue("{1:10}"))

# Wide values can be truncated
collapse(glue("{1:10}"), width = 5)

collapse(1:4, ",", last = " and ")
#> 1, 2, 3 and 4
}
