% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate R code}
\usage{
evaluate(code, envir, data)
}
\arguments{
\item{code}{R code to evaluate}

\item{envir}{environment to evaluate the code in}

\item{data}{listish object to evaluate the code in}
}
\description{
This is a simple wrapper around \code{eval(parse())} which provides a more
consistent interface than the default functions.
If \code{data} is \code{NULL} then the code is evaluated in the environment. If \code{data}
is not \code{NULL} than the code is evaluated in the \code{data} object first, with
the enclosing environment of \code{envir}.
}
\details{
This function is designed to be used within transformers to evaluate the
code in the glue block.
}
