% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_glue_expr}
\alias{extract_glue_expr}
\title{Extract all expressions to be evaluated by \code{glue()}}
\usage{
extract_glue_expr(...)
}
\arguments{
\item{...}{passed to \code{glue()}}
}
\value{
A character vector of expressions to be evaluated
}
\description{
Extract all expressions to be evaluated by \code{glue()}
}
\examples{
extract_glue_expr("foo {character(0)} {bar}")
}
