% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstrings.R
\name{fstring}
\alias{fstring}
\alias{f}
\title{Format a string}
\usage{
fstring(x, sep = "", fun = as.character, envir = parent.frame())

f(x, sep = "", fun = as.character, envir = parent.frame())
}
\arguments{
\item{x}{the string to format}

\item{sep}{separator used to collapse elements if \code{fun} returns more than one item.}

\item{fun}{Function to call to format each expression.}

\item{envir}{Environment to evaluate each expression in. Expressions are
evaluated in the order they appear.}
}
\description{
Format a string
}
