% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{collapse}
\alias{collapse}
\alias{glue-deprecated}
\title{Deprecated Functions}
\usage{
collapse(x, sep = "", width = Inf, last = "")
}
\description{
These functions are Deprecated in this release of glue, they will be removed
in a future version.
}
\keyword{internal}
