% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue-package.R
\docType{package}
\name{glue-package}
\alias{glue-package}
\title{glue: Interpreted String Literals}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An implementation of interpreted string literals, inspired by Python's Literal String Interpolation \url{https://www.python.org/dev/peps/pep-0498/} and Docstrings \url{https://www.python.org/dev/peps/pep-0257/} and Julia's Triple-Quoted String Literals \url{https://docs.julialang.org/en/v1.3/manual/strings/#Triple-Quoted-String-Literals-1}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glue.tidyverse.org/}
  \item \url{https://github.com/tidyverse/glue}
  \item Report bugs at \url{https://github.com/tidyverse/glue/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@posit.co} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

Authors:
\itemize{
  \item Jennifer Bryan \email{jenny@posit.co} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})
  \item Jim Hester (\href{https://orcid.org/0000-0002-2739-7082}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
