% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs.R
\name{glue-vctrs}
\alias{glue-vctrs}
\alias{vec_ptype2.glue}
\alias{vec_cast.glue}
\title{vctrs methods for glue}
\usage{
vec_ptype2.glue(x, y, ...)

vec_cast.glue(x, to, ...)
}
\description{
vctrs methods for glue
}
\keyword{internal}
