% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_batch_get_partition}
\alias{glue_batch_get_partition}
\title{Retrieves partitions in a batch request}
\usage{
glue_batch_get_partition(CatalogId, DatabaseName, TableName,
  PartitionsToGet)
}
\arguments{
\item{CatalogId}{The ID of the Data Catalog where the partitions in question reside. If
none is supplied, the AWS account ID is used by default.}

\item{DatabaseName}{[required] The name of the catalog database where the partitions reside.}

\item{TableName}{[required] The name of the partitions' table.}

\item{PartitionsToGet}{[required] A list of partition values identifying the partitions to retrieve.}
}
\description{
Retrieves partitions in a batch request.
}
\section{Request syntax}{
\preformatted{glue$batch_get_partition(
  CatalogId = "string",
  DatabaseName = "string",
  TableName = "string",
  PartitionsToGet = list(
    list(
      Values = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
