% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_resource_policy}
\alias{glue_delete_resource_policy}
\title{Deletes a specified policy}
\usage{
glue_delete_resource_policy(PolicyHashCondition)
}
\arguments{
\item{PolicyHashCondition}{The hash value returned when this policy was set.}
}
\description{
Deletes a specified policy.
}
\section{Request syntax}{
\preformatted{glue$delete_resource_policy(
  PolicyHashCondition = "string"
)
}
}

\keyword{internal}
