% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_delete_trigger}
\alias{glue_delete_trigger}
\title{Deletes a specified trigger}
\usage{
glue_delete_trigger(Name)
}
\arguments{
\item{Name}{[required] The name of the trigger to delete.}
}
\description{
Deletes a specified trigger. If the trigger is not found, no exception
is thrown.
}
\section{Request syntax}{
\preformatted{glue$delete_trigger(
  Name = "string"
)
}
}

\keyword{internal}
