% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_jobs}
\alias{glue_get_jobs}
\title{Retrieves all current job definitions}
\usage{
glue_get_jobs(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{A continuation token, if this is a continuation call.}

\item{MaxResults}{The maximum size of the response.}
}
\description{
Retrieves all current job definitions.
}
\section{Request syntax}{
\preformatted{glue$get_jobs(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
