% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_operations.R
\name{glue_get_resource_policy}
\alias{glue_get_resource_policy}
\title{Retrieves a specified resource policy}
\usage{
glue_get_resource_policy()
}
\description{
Retrieves a specified resource policy.
}
\section{Request syntax}{
\preformatted{glue$get_resource_policy()
}
}

\keyword{internal}
