% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe.R
\name{glue_safe}
\alias{glue_safe}
\alias{glue_data_safe}
\title{Safely interpolate strings}
\usage{
glue_safe(..., .envir = parent.frame())

glue_data_safe(.x, ..., .envir = parent.frame())
}
\description{
\code{glue_safe()} and \code{glue_data_safe()} differ from \code{glue()} and \code{glue_data()}
in that the safe versions only look up symbols from an environment use
\code{\link[=get]{get()}} they do not execute any R code. This makes them suitable when used
with untrusted input, such as inputs in a shiny application, where using the
normal functions would allow an attacker to execute arbitrary code.
}
