% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{identity_transformer}
\alias{identity_transformer}
\title{Parse and Evaluate R code}
\usage{
identity_transformer(text, envir = parent.frame())
}
\arguments{
\item{text}{Text (typically) R code to parse and evaluate.}

\item{envir}{environment to evaluate the code in}
}
\description{
This is a simple wrapper around \code{eval(parse())}, used as the default
transformer.
}
\seealso{
\code{vignette("transformers", "glue")} for documentation on creating
custom glue transformers and some common use cases.
}
