% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoting.R
\name{quoting}
\alias{quoting}
\alias{single_quote}
\alias{double_quote}
\alias{backtick}
\title{Quoting operators}
\usage{
single_quote(x)

double_quote(x)

backtick(x)
}
\arguments{
\item{x}{A character to quote.}
}
\description{
These functions make it easy to quote each individual element and are useful
in conjunction with \code{\link[=glue_collapse]{glue_collapse()}}.
}
\examples{
x <- 1:5
glue('Values of x: {glue_collapse(backtick(x), sep = ", ", last = " and ")}')
}
