% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstrings.R
\name{collapse}
\alias{collapse}
\title{Collapse a character vector}
\usage{
collapse(x, sep = "")
}
\arguments{
\item{x}{The character vector to collapse.}

\item{sep}{a character string to separate the terms.  Not
    \code{\link{NA_character_}}.}
}
\description{
Collapses a character vector of any length into a length 1 vector.
}
\examples{
collapse(f("{1:10}"))
}
