% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstrings.R
\name{fstring}
\alias{fstring}
\alias{fstring_}
\alias{f_}
\alias{fstring}
\alias{f}
\title{Format and interpolate a string}
\usage{
fstring_(.x, ..., .sep = "", .envir = parent.frame(), .fun = as.character)

f_(.x, ..., .sep = "", .envir = parent.frame(), .fun = as.character)

fstring(..., .sep = "", .envir = parent.frame(), .fun = as.character)

f(..., .sep = "", .envir = parent.frame(), .fun = as.character)
}
\arguments{
\item{.x}{An environment, list or data frame.}

\item{...}{String(s) to format, multiple inputs are concatenated together before formatting.}

\item{.sep}{Separator used to separate elements.}

\item{.envir}{Environment to evaluate each expression in. Expressions are
evaluated from left to right.}

\item{.fun}{Function to used to format each result.}
}
\description{
Expressions enclosed by braces will be evaluated as R code. Single braces
can be inserted by doubling them. The inputs are not vectorized.
}
\examples{
name <- "Fred"
age <- 50
anniversary <- as.Date("1991-10-12")
f('My name is {name},',
  'my age next year is {age + 1},',
  'my anniversary is {format(anniversary, "\%A, \%B \%d, \%Y")}.')

# single braces can be inserted by doubling them
f("My name is {name}, not {{name}}.")

# Named arguments can also be supplied
f('My name is {name},',
  ' my age next year is {age + 1},',
  ' my anniversary is {format(anniversary, "\%A, \%B \%d, \%Y")}.',
  name = "Joe",
  age = 40,
  anniversary = as.Date("2001-10-12"))
}
\seealso{
\url{https://www.python.org/dev/peps/pep-0498/} upon which this is based.
}
