% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstrings.R
\name{fstring}
\alias{fstring}
\alias{f}
\title{Intrepret string literals}
\usage{
fstring(..., .sep = "", .fun = as.character, .envir = parent.frame())

f(..., .sep = "", .fun = as.character, .envir = parent.frame())
}
\arguments{
\item{...}{the string(s) to format, multiple inputs are concatonated together before formatting.}

\item{.sep}{separator used to collapse elements if \code{fun} returns more than one result.}

\item{.fun}{Function to call to format each expression.}

\item{.envir}{Environment to evaluate each expression in. Expressions are
evaluated in the order they appear.}
}
\description{
Intrepret string literals
}
\examples{
name <- "Fred"
age <- 50
anniversary <- as.Date("1991-10-12")
f('My name is {name},',
   'my age next year is {age + 1},',
   'my anniversary is {format(anniversary, "\%A, \%B \%d, \%Y")}.')

# Can also use alternative f$'text' syntax with `+`
f$'My name is {name},' +
  'my age next year is {age + 1},' +
  'my anniversary is {format(anniversary, "\%A, \%B \%d, \%Y")}'
}
